Update the classifier tool for multi-class classification
After re-implementing the Classifier class, the classification
example (now more of a tool) can be used for classifying one
or multiple images contained in a folder.

Additional function are provided for preprocessing, reading
input and getting the gradient of the k-th classifier of the
network with respect to the input.

More functionality can be added as needed as the class is more
general than before.